add a rake task to send sign-out reminders via heroku scheduler
The scheduled task created to run this rake task should run once an hour. Heroku
scheduler is capable of executing tasks at a particular hour and minute once per
day, but the time must be specified in UTC, so we'll use the same "early exit"
strategy we used in the weekly report. In that case, we used a daily scheduled
task to implement a weekly task, whereas in this case, we use an hourly
scheduled task scheduled in UTC to implement a daily task scheduled in an
arbitrary time zone.